Update translate.cpp for Prusa M555
Process Prusa-specific Gcode M555, which sets "active bed area" for G29 bed probing.  For a single shift/copy, only the X and Y parameters are translated.  For a multiple shift/copy, the combination of all new objects is used to set the new X,Y orgin point as well as the new Width and Height (W, H) parameters.